+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 0.1.3 "Vaticano" (08.05.2015 - ...05.2015)
+* Check for null values for SEACRH
+* ORDER BY for SEARCH operator
+
 ### 0.1.2 "Firenze" (06.05.2015 - 07.05.2015)
 * Simple compilation of SEARCH operator
 * SUM(),COUNT(),MIN(),MAX(),FIRST(),LAST() search aggregators